Adding mesh geometry
